css changes

+++ Models/Loan.cs
@@ -13,12 +13,13 @@ namespace LoanSharkMVC.Models
 
         public int Term { get; set; }
 
-        public int Payment { get; set; }
+        public decimal Payment { get; set; }
 
         public decimal TotalInterest { get; set; }
 
         public decimal TotalCost { get; set; }
 
         public List<LoanPayment> Payments { get; set; } = new List<LoanPayment>();
+
     }
 }
